% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMap.R
\name{CellMap}
\alias{CellMap}
\title{CellMap}
\usage{
CellMap(
  st.obj = st.obj,
  sc.obj = sc.obj,
  coord = c("x", "y"),
  resolution = 0.3,
  celltype.column = "idents",
  sc.sub.size = NULL,
  min.sc.cell = 50,
  factor.size = 0.1,
  seed.num = 10,
  pvalue.cut = 0.1,
  knn = 5,
  mean.cell.num = 5,
  max.cell.num = 10,
  n.workers = 4,
  verbose = TRUE
)
}
\arguments{
\item{st.obj}{Seurat object of spatial transcriptome data.}

\item{sc.obj}{Seurat object of scRNA-seq data.}

\item{coord}{Coordinates column names in ST images slot.coord = c("x","y") or coord = c("imagerow","imagecol").}

\item{resolution}{The clustering resolution for spatial transcriptomics data. Default:0.3.}

\item{celltype.column}{The column name for cell type in the single-cell Seurat object, with the default value as "idents".}

\item{sc.sub.size}{Downsampling proportion or number for scRNA-seq data. Default: NULL.}

\item{factor.size}{Factor size for scaling the weight of gene expression. Default: 0.1.}

\item{seed.num}{Number of seed genes of each cell type for recognizing candidate markers. Default: 10.}

\item{pvalue.cut}{Threshold for filtering cell type marker genes. Default: 0.1.}

\item{knn}{The number of nearest neighboring single cells for each spot. Set to 5 for low-resolution data and 1 for high-resolution data.}

\item{mean.cell.num}{The average number of single cells in the spot.Set to 5 for low-resolution data and 1 for high-resolution data.}

\item{max.cell.num}{The maximum number of cells within each spot, if equal to 1, indicates that each spot contains only a single cell.}

\item{n.workers}{Number of cores to be used for parallel processing. Default: 4.}

\item{verbose}{Show running messages or not. Default: TRUE.}

\item{min.sc.cells}{The minimum number of cell types in scRNA-seq data.Default: 50}
}
\description{
CellMap
}
