% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellmapOne.R
\name{cellmapOne}
\alias{cellmapOne}
\title{Estimate the cell type proportions of mixture bulk RNA}
\usage{
cellmapOne(
  strBulk,
  strSC,
  strPrefix = substr(strBulk, 1, nchar(strBulk) - 4),
  method = "CellMap",
  geneNameReady = F,
  ensemblPath = "Data/",
  ensemblV = 97,
  expCutoff = 1e+06,
  cellMapDEG = "DESeq2",
  cellMapGeneN = 100,
  core = 2
)
}
\arguments{
\item{strBulk}{The full path to the query mixture bulk expression file.
Expression matrix separated by tabs with rows are genes, columns are samples.
First row includes the sample names or ids, while first column consists of gene symbols or gene ensembl id.}

\item{strSC}{The paths to the expression matrix of a sc/sn RNAseq datasets (.rds).
The expression matrix with rows are genes (official gene symbol is required as first column);
and columns are cells with cell type and sample information encoded in to the column names (cellType|sampleID|…).
The sample information will be used by MuSiC.}

\item{strPrefix}{The prefix with path of the result files.
There are two files produced: a pdf file contains all cell type decomposition figures;
a tab separated table file including composition and p-values.}

\item{method}{'CellMap' or 'MuSiC' to be choosen for decomposition. Default is CellMap.}

\item{geneNameReady}{A boolean to indicate if the gene names in the query mixture bulk expression matrix
is official symbol already.
The \code{FALSE} option also works with the official symbol is used in the expression matrix.
Default is \code{FALSE}, which enable to find official symbol by an R package called \code{biomaRt}.}

\item{ensemblPath}{The path to a folder where ensembl gene definition is/will be saved.
The ensembl gene definition file will be saved if it never run before.
Default is \emph{Data/} in the current working directory.}

\item{ensemblV}{The version of the ensembl to be used for the input query bulk expression. Default is 97.}

\item{expCutoff}{The minimal total measure of a cell type to be considered. Default is 1M}

\item{cellMapDEG}{One from ‘edgeR’, ‘DESeq2’, ‘voom’ or ‘Top’ can be chosen.
This indicates the method for identifying the cell type signature genes.
Since it is from one dataset, and no batch correction, default is DESeq2.}

\item{cellMapGeneN}{A numeric indicates the maximin number of signature genes for each cell type in an iteration.
Default is 100.}

\item{core}{The number of computation nodes could be used. Default is 2.}
}
\description{
This function estimates cell type proportions of mixture bulk RNA samples based on one sc/sn RNAseq expression matrix.
}
\examples{
strBulk <- system.file("extdata","bulk.txt",package="cellmap")
strSC <- system.file("extdata","GSE103723.rds",package="cellmap")
cellmapOne(strBulk,strSC,strPrefix="~/bulk_decomp",core=16)
cellmapOne(strBulk,strSC,strPrefix="~/bulk_decomp",method="MuSiC",core=16)


}
