% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMapper.R
\name{CellMapper}
\alias{CellMapper}
\title{CellMapper}
\usage{
CellMapper(
  st.data = st.data,
  coord.df = coord.df,
  ref.expr = ref.expr,
  ref.anno = ref.anno,
  factor.size = 0.1,
  seed.num = 10,
  pvalue.cut = 0.1,
  knn = 5,
  verbose = TRUE
)
}
\arguments{
\item{st.data}{st data Seurat object.}

\item{coord.df}{The spatial coordinates of spatial transcriptomics images.}

\item{ref.expr}{Single-cell gene expression profile.}

\item{ref.anno}{Cell type information of gene expression profile, corresponding to the above \code{ref.expr}.}

\item{factor.size}{Factor size for scaling the weight of gene expression. Default: 0.1.}

\item{seed.num}{Number of seed genes of each cell type for recognizing candidate markers. Default: 10.}

\item{pvalue.cut}{Threshold for filtering cell type marker genes. Default: 0.1}

\item{knn}{The number of nearest neighboring single cells for each spot. Default: 0.1}

\item{verbose}{Show running messages or not. Default: TRUE.}
}
\description{
CellMapper
}
