% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputParams.r
\name{checkInputParams}
\alias{checkInputParams}
\title{checkInputParams}
\usage{
checkInputParams(
  st.obj,
  sc.obj,
  coord,
  celltype.column,
  sc.sub.size,
  min.sc.cell,
  factor.size,
  seed.num,
  pvalue.cut,
  knn,
  mean.cell.num,
  max.cell.num,
  n.workers,
  verbose
)
}
\value{
TRUE.
}
\description{
checkInputParams
}
