% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createSeuratObj}
\alias{createSeuratObj}
\title{createSeuratObj}
\usage{
createSeuratObj(st.obj, sc.obj, mapping)
}
\arguments{
\item{st.obj}{Seurat object of Spatial Transcriptome data.}

\item{sc.obj}{Seurat object of single cell data.}

\item{mapping}{single cell mapping spatial results.}
}
\value{
Seurat object.
}
\description{
Create a Seurat object.
}
\examples{
sc.out <- createSeuratObj(st.obj,sc.obj, mapping)
}
