% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createSpObj}
\alias{createSpObj}
\title{createSpatialObject}
\usage{
createSpObj(
  counts,
  coord.df,
  coord.label = c("x", "y"),
  meta.data = NULL,
  class = "SlideSeq"
)
}
\arguments{
\item{counts}{A matrix of UMI counts, where rows represent genes and columns represent barcodes.}

\item{coord.df}{A data frame containing spatial coordinates for the barcodes. The coordinates should be in the first two columns.}

\item{coord.label}{A character vector specifying the column names in coord.df for the spatial coordinates. Default is c("x", "y").}

\item{meta.data}{Optional metadata data frame, where rows are barcodes.}

\item{class}{A character string specifying the class name for the Spatial image slot. Default is "SlideSeq".}
}
\value{
A Seurat object for spatial transcriptomics data with the spatial image slot populated.
}
\description{
Creates a Spatial Seurat object for spatial transcriptomics (ST) data from various platforms.
This function ensures that only matching barcodes between the count matrix and coordinates are included.
}
