% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{getCellClusterInfo}
\alias{getCellClusterInfo}
\title{getCellClusterInfo}
\usage{
getCellClusterInfo(nearCells, st.obj)
}
\arguments{
\item{nearCells}{A list where the names are spot identifiers and the values are cell identifiers.}

\item{st.obj}{Seurat object of spatial transcriptomics data with metadata including cluster information.}
}
\description{
This function retrieves the cluster information for each cell based on the spot-cell correspondence.
}
