% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCellNumber}
\alias{getCellNumber}
\title{getCellNumber}
\usage{
getCellNumber(ST.data, mean.cell.num = 5)
}
\arguments{
\item{ST.data}{The ST expression data has genes as rows and spots as columns.}

\item{mean.cell.num}{Mean cell number of spot,for 10X visium data,default:5.}
}
\value{
The number of cells of each spot.
}
\description{
Estimate the number of cells in each spot using linear regression.
}
