% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{getCells}
\alias{getCells}
\title{getCells}
\usage{
getCells(sc.st.dist, k = knn)
}
\arguments{
\item{sc.st.dist}{A distance matrix of single cell and spot.}

\item{k}{Number of nearest neighboring single cells for each spot.}
}
\value{
A list of nearest single cells corresponding to the spot.
}
\description{
Obtaining the k nearest single cells of each spot.
}
