% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{getDistMatrix}
\alias{getDistMatrix}
\title{getDistMatrix}
\usage{
getDistMatrix(sc.st.obj)
}
\arguments{
\item{sc.st.obj}{Seurat object of integrated sc and st data.}
}
\value{
A distance matrix of single cell and spot.
}
\description{
Calculate the distance matrix between single cells and spatial spots.
}
