% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getELDist}
\alias{getELDist}
\title{getELDist}
\usage{
getELDist(scRNA.data, ST.data)
}
\arguments{
\item{scRNA.data}{scRNA-seq data of sampled cells.}

\item{ST.data}{ST data.}
}
\value{
Euclidean distance of single cells and spots.
}
\description{
Calculate the Euclidean distance between single cells and spots.
}
