% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{getInterdata}
\alias{getInterdata}
\title{getInterdata}
\usage{
getInterdata(sc.obj, st.obj, coord = c("imagerow", "imagecol"), markers)
}
\arguments{
\item{sc.obj}{sc.obj Seurat object of scRNA-seq data.}

\item{st.obj}{Seurat object of spatial transcriptome data.}

\item{coord}{Coordinates column names in st images slot.}

\item{markers}{Feature genes for intergration.}
}
\value{
Seurat object of integrated sc and st data.
}
\description{
Integrate single-cell(sc) and spatial transcriptomic(st) data.
}
