% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{getNearCells}
\alias{getNearCells}
\title{getNearCells}
\usage{
getNearCells(sim, num.cells)
}
\arguments{
\item{sim}{Similarity matrix of single cells and spots.}

\item{num.cells}{The number of single cells contained in each spot.}
}
\value{
A list of single cells corresponding to spots.
}
\description{
Compute the nearest single cells for each spot.
}
