% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNormalizeData}
\alias{getNormalizeData}
\title{getNormalizeData}
\usage{
getNormalizeData(data)
}
\arguments{
\item{data}{Data that needs to be normalized.}
}
\value{
Normalized expression data.
}
\description{
Data normalization.
}
