% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRandomCoords}
\alias{getRandomCoords}
\title{getRandomCoords}
\usage{
getRandomCoords(st.obj, sim.cells, n.workers = 4)
}
\arguments{
\item{st.obj}{Seurat object of spatial transcriptome data.}

\item{sim.cells}{A list of single cells corresponding to spots.}

\item{n.workers}{Number of cores to be used for parallel processing. Default: 4.}

\item{spot.coords}{The spatial coordinates of the spot, including two columns, imagerow and imagecol.}
}
\value{
Spatial coordinates of simulated single cells.
}
\description{
Get random coordinates for mapping single-cells to spatial.
}
\examples{
cell.coords <- getRandomCoords(st.obj,sim.cells,n.workers = 4)
}
