% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSimCells}
\alias{getSimCells}
\title{getSimCells}
\usage{
getSimCells(sim.matrix, num.cells)
}
\arguments{
\item{sim.matrix}{Similarity matrix of single cell and spatial spot,row:spot,column:single cell.}

\item{num.cells}{The number of cells of each spot.}
}
\value{
A list of cells corresponding to each spot.
}
\description{
Get the most similar single cell for each spot.
}
