% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{linearAllocation}
\alias{linearAllocation}
\title{linearAllocation}
\usage{
linearAllocation(sim, num.cells)
}
\arguments{
\item{sim}{Similarity matrix of single cells and spots.}

\item{num.cells}{The number of single cells contained in each spot.}
}
\value{
A list of single cells corresponding to spots.
}
\description{
Apply linear allocation algorithm to achieve one-to-one mapping of single cells to spots.
}
