% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapSctoSpatial}
\alias{mapSctoSpatial}
\title{mapSctoSpatial}
\usage{
mapSctoSpatial(st.obj, sc.obj, sim.cells, coord.df)
}
\arguments{
\item{st.obj}{Seurat object of Spatial Transcriptome data.}

\item{sc.obj}{Seurat object of single cells data.}

\item{sim.cells}{A list of cells corresponding to each spot.}
}
\value{
Mapping results for single cells and spatial coordinates.
}
\description{
Assign single cells to spatial coordinates.
}
