% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multipleProcess}
\alias{multipleProcess}
\title{Set Up Parallel Processing}
\usage{
multipleProcess(n.workers = 4, max.size = 40000)
}
\arguments{
\item{n.workers}{Number of cores to be used for parallel processing. Default: 8.}

\item{max.size}{Maximum size of global objects allowed in the future framework, in megabytes (MB). Default: 40,000 MB.}
}
\description{
Configures and opens multiple workers for Seurat or other parallel processing in R.
}
