% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processScData}
\alias{processScData}
\title{Process scRNA-seq data.}
\usage{
processScData(
  sc.obj,
  celltype.column = "idents",
  sc.sub.size = NULL,
  min.sc.cells = 50
)
}
\arguments{
\item{sc.obj}{Seurat object of scRNA-seq data.}

\item{celltype.column}{The column name for cell type in the single-cell Seurat object, with the default value as "idents".}

\item{sc.sub.size}{Downsampling proportion or number for scRNA-seq data. Default: NULL}

\item{min.sc.cells}{The minimum number of cell types in scRNA-seq data.Default: 50}
}
\value{
A Seurat object with SCTransform normalization.
}
\description{
A function to perform SCTransform normalization, PCA, UMAP, neighbor finding, and clustering on a Seurat object for spatial transcriptomics data.
}
\examples{
sc.obj <- processScData(sc.obj)
}
