% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processSpatialData}
\alias{processSpatialData}
\title{Process Spatial Transcriptomics Data}
\usage{
processSpatialData(
  st.obj,
  assay = "Spatial",
  pca_dims = 30,
  resolution = resolution
)
}
\arguments{
\item{st.obj}{Seurat object of spatial transcriptome data.}

\item{assay}{The assay to be used for SCTransform normalization. Default is "Spatial".}

\item{pca_dims}{Number of principal components to use for UMAP and neighbor finding. Default is 30.}

\item{resolution}{The resolution parameter for the FindClusters function. Default is 0.3.}
}
\value{
A Seurat object with SCTransform normalization, PCA, UMAP, and clustering results.
}
\description{
A function to perform SCTransform normalization, PCA, UMAP, neighbor finding, and clustering on a Seurat object for spatial transcriptomics data.
}
\examples{
st.obj <- processSpatialData(st.obj,resolution = resolution)
}
