% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{trainModel}
\alias{trainModel}
\title{trainModel}
\usage{
trainModel(st.data, sc.data, nearCells, markers)
}
\arguments{
\item{st.data}{st data Seurat object.}

\item{sc.data}{sc data Seurat object.}

\item{nearCells}{A list of nearest single cells corresponding to the spot.}

\item{sig.mat}{Marker gene expression matrix.}
}
\value{
A list of predicted results.
}
\description{
Train a random forest model and make predictions.
}
