% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCellNumber}
\alias{getCellNumber}
\title{getCellNumber}
\usage{
getCellNumber(st.data, mean.cell.num = mean.cell.num)
}
\arguments{
\item{st.data}{The spatial transcriptome expression data has genes as rows and spots as columns.}

\item{mean.cell.num}{Mean cell number of spot;set to 5 for 10X Visium data; set to 1 for high-resolution data.}
}
\value{
The number of cells of each spot.
}
\description{
Estimate the number of cells in each spot using linear regression.
}
