% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getELDist}
\alias{getELDist}
\title{getELDist}
\usage{
getELDist(sc.data, st.data)
}
\arguments{
\item{sc.data}{scRNA-seq data of sampled cells.}

\item{st.data}{Spatial transcriptomics expression data.}
}
\value{
Euclidean distance of single cells and spots.
}
\description{
Calculate the Euclidean distance between single cells and spots.
}
