% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{getInterdata}
\alias{getInterdata}
\title{getInterdata}
\usage{
getInterdata(sc.data, st.data, coord = c("imagerow", "imagecol"), markers)
}
\arguments{
\item{sc.data}{sc data Seurat object.}

\item{st.data}{st data Seurat object.}

\item{coord}{coordinates column names in st images slot.}

\item{nfeatures}{number of feature genes for intergration.}
}
\value{
Seurat object of integrated sc and st data.
}
\description{
Integrate single-cell(sc) and spatial transcriptomic(st) data.
}
