% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRandomCoords}
\alias{getRandomCoords}
\title{getRandomCoords}
\usage{
getRandomCoords(spot.coords, sim.cells)
}
\arguments{
\item{spot.coords}{The spatial coordinates of the spot, including two columns, imagerow and imagecol.}

\item{sim.cells}{A list of single cells corresponding to spots.}
}
\value{
Spatial coordinates of simulated single cells.
}
\description{
Get random coordinates for mapping single-cells to spatial.
}
\examples{
cell.coords <- getRandomCoords(coord.df,sim.cells)
}
