% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{getSimMatrix}
\alias{getSimMatrix}
\title{getSimMatrix}
\usage{
getSimMatrix(pred, st.data, sc.data, markers)
}
\arguments{
\item{pred}{A list of predicted results.}

\item{st.data}{st data Seurat object.}

\item{sc.data}{sc data Seurat object.}

\item{sig.mat}{Marker gene expression matrix.}
}
\value{
A similarity matrix of single cells and spots.
}
\description{
Calculate the similarity of single cells and spots.
}
