% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{trainModel}
\alias{trainModel}
\title{trainModel}
\usage{
trainModel(st.obj, sc.obj, nearCells, markers, n.workers = 4)
}
\arguments{
\item{st.obj}{Seurat object of spatial transcriptome data.}

\item{sc.obj}{Seurat object of scRNA-seq data.}

\item{nearCells}{A list of nearest single cells corresponding to the spot.}

\item{markers}{Cell type specific marker genes.}

\item{n.workers}{Number of cores to be used for parallel processing. Default: 4.}
}
\value{
A list of predicted results.
}
\description{
Train a random forest model and make predictions.
}
